% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filterSets.R
\name{filterSets}
\alias{filterSets}
\title{prediction}
\usage{
filterSets(df, hyp)
}
\arguments{
\item{df:}{dataframe that contains all the data \code{inputParameter1}}

\item{hyp:}{choice of hypothesis\code{inputParameter2}}
}
\value{
df: filtered data frame
}
\description{
This filters data frame wit hthe hypothesis by you choice
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

