% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitAndPredict.R
\name{fitAndPredict}
\alias{fitAndPredict}
\title{prediction}
\usage{
fitAndPredict(base, dataTable, predictTable, model)
}
\arguments{
\item{base:}{reference base at the alignming position\code{inputParameter1}}

\item{datatable:}{training dataset contains all the data \code{inputParameter2}}

\item{predictTable:}{new dataset to be predicted \code{inputParameter2}}

\item{model:}{the fitted model to be used \code{inputParameter2}}
}
\value{
df: return original data fram with a new column with the predicted class
}
\description{
This function is for modeling the distribution of errors
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

