% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modeling.R
\name{modeling}
\alias{modeling}
\title{modeling}
\usage{
modeling(base, df, model)
}
\arguments{
\item{base:}{reference base at the alignming position\code{inputParameter1}}

\item{df:}{dataframe that contains all the data \code{inputParameter2}}

\item{model:}{the model to be used \code{inputParameter2}}
}
\value{
y1: return fitted model or the only training class
}
\description{
This function is for modeling the distribution of errors
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

