% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predicting.R
\name{predicting}
\alias{predicting}
\title{prediction}
\usage{
predicting(base, model, df)
}
\arguments{
\item{base:}{reference base at the alignming position\code{inputParameter1}}

\item{df:}{dataframe that contains all the data \code{inputParameter2}}

\item{model:}{the fitted model to be used \code{inputParameter2}}
}
\value{
df: return original data fram with a new column with the predicted class
}
\description{
This function is for modeling the distribution of errors
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

