% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tgirthamr.R
\name{tgirthamr}
\alias{tgirthamr}
\title{prediction}
\usage{
tgirthamr(predictTable, model, enzyme, seqErr, pCutOff, resultFile, hyp, dbpath,
  devMode)
}
\arguments{
\item{predictTable:}{data table that store mismatch of the library\code{inputParameter1}}

\item{model:}{model of your choice  (knn/rf/..)\code{inputParameter2}}

\item{enzyme:}{what enzyme did you used for preparing the library?  \code{inputParameter3}}

\item{seqErr:}{theoretical sequencin/PCR error  \code{inputParameter4}}

\item{pCutOff:}{false discovery rate threshold  \code{inputParameter5}}

\item{hyp:}{hypothesis to use  \code{inputParameter6}}

\item{dbpath:}{the path storing the database  \code{inputParameter7}}

\item{devMode:}{1 or 0 for outputing data that with extensive information or not \code{inputParameter8}}
}
\value{
df: return original data fram with a new column with the predicted class
}
\description{
This function is for modeling the distribution of errors
}
\examples{
R code here showing how your function works
}
\keyword{keywords}

